[Site] Update demo components paths
As mentionned in #2333, demo component classes did not follow current "default" configuration, leading to paths not in line with documentation.

Easier than anticipated to fix.
