--- ux.symfony.com/src/Twig/Alert.php
@@ -1,29 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace App\Twig;
-
-use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
-
-#[AsTwigComponent]
-class Alert
-{
-    public string $type = 'success';
-    public string $message;
-
-    public function getIcon(): string
-    {
-        return match ($this->type) {
-            'success' => 'bi:check-circle',
-            'danger' => 'bi:exclamation-circle',
-        };
-    }
-}
