+++ ux.symfony.com/src/Twig/Components/HomepageTerminalSwapper.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Twig;
+namespace App\Twig\Components;
 
 use App\Model\UxPackage;
 use App\Service\UxPackageRepository;
