+++ ux.symfony.com/src/Twig/Components/InvoiceCreatorLineItem.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Twig;
+namespace App\Twig\Components;
 
 use App\Entity\Product;
 use App\Repository\ProductRepository;
