+++ ux.symfony.com/src/Twig/Components/MealPlanner.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Twig;
+namespace App\Twig\Components;
 
 use App\Form\MealPlannerForm;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
