+++ ux.symfony.com/src/Twig/Components/ProductGrid.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Twig;
+namespace App\Twig\Components;
 
 use App\Service\EmojiCollection;
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
@@ -17,15 +17,14 @@ use Symfony\UX\LiveComponent\Attribute\LiveAction;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\ComponentToolsTrait;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
-use Symfony\UX\TwigComponent\Attribute\ExposeInTemplate;
 
-#[AsLiveComponent('ProductGrid2')]
-class ProductGrid2
+#[AsLiveComponent('ProductGrid')]
+class ProductGrid
 {
     use ComponentToolsTrait;
     use DefaultActionTrait;
 
-    private const PER_PAGE = 9;
+    private const PER_PAGE = 10;
 
     #[LiveProp]
     public int $page = 1;
@@ -34,24 +33,10 @@ class ProductGrid2
     {
     }
 
-    public function getItems(): array
+    #[LiveAction]
+    public function more(): void
     {
-        $items = [];
-        $emojis = $this->emojis->paginate($this->page, self::PER_PAGE);
-        foreach ($emojis as $i => $emoji) {
-            $items[] = [
-                'id' => ($this->page - 1) * self::PER_PAGE + $i,
-                'emoji' => $emoji,
-            ];
-        }
-
-        return $items;
-    }
-
-    #[ExposeInTemplate('per_page')]
-    public function getPerPage(): int
-    {
-        return self::PER_PAGE;
+        ++$this->page;
     }
 
     public function hasMore(): bool
@@ -59,9 +44,31 @@ class ProductGrid2
         return \count($this->emojis) > ($this->page * self::PER_PAGE);
     }
 
-    #[LiveAction]
-    public function more(): void
+    public function getItems(): array
     {
-        ++$this->page;
+        $emojis = $this->emojis->paginate($this->page, self::PER_PAGE);
+        $colors = $this->getColors();
+
+        $items = [];
+        foreach ($emojis as $i => $emoji) {
+            $items[] = [
+                'id' => $id = ($this->page - 1) * self::PER_PAGE + $i,
+                'emoji' => $emoji,
+                'color' => $colors[$id % \count($colors)],
+            ];
+        }
+
+        return $items;
+    }
+
+    public function getColors(): array
+    {
+        return [
+            '#fbf8cc', '#fde4cf', '#ffcfd2',
+            '#f1c0e8', '#cfbaf0', '#a3c4f3',
+            '#90dbf4', '#8eecf5', '#98f5e1',
+            '#b9fbc0', '#b9fbc0', '#ffc9c9',
+            '#d7ffc9', '#c9fffb',
+        ];
     }
 }
