+++ ux.symfony.com/src/Twig/Components/TodoListForm.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Twig;
+namespace App\Twig\Components;
 
 use App\Entity\TodoList;
 use App\Form\TodoListFormType;
