+++ ux.symfony.com/templates/demos/live_component/auto_validating_form.html.twig
@@ -1,7 +1,7 @@
 {% extends 'demos/live_demo.html.twig' %}
 
 {% block code_block_left %}
-    <twig:CodeBlock filename="src/Twig/RegistrationForm.php" height="400px" />
+    <twig:CodeBlock filename="src/Twig/Components/RegistrationForm.php" height="400px" />
 {% endblock %}
 
 {% block code_block_right %}
