+++ ux.symfony.com/templates/demos/live_component/dependent_form_fields.html.twig
@@ -2,7 +2,7 @@
 
 {% block code_block_full %}
     <twig:Code:TabbedCodeBlocks :files="[
-        'src/Twig/MealPlanner.php',
+        'src/Twig/Components/MealPlanner.php',
         'templates/components/MealPlanner.html.twig',
         'src/Form/MealPlannerForm.php',
     ]" />
