+++ ux.symfony.com/templates/demos/live_component/infinite_scroll.html.twig
@@ -26,7 +26,7 @@
 
     <section class="py-5" style="padding-top: 6rem;">
         <div class="container container-xxl">
-            <twig:Code:CodeWithExplanationRow filename="src/Twig/ProductGrid.php" reversed sticky style="--bs-gutter-x: 4rem;">
+            <twig:Code:CodeWithExplanationRow filename="src/Twig/Components/ProductGrid.php" reversed sticky style="--bs-gutter-x: 4rem;">
 This component is quite standard: the page number as a `LiveProp`, a `LiveAction`
 to load the next page, and a `getItems` method to retrieve the page results.
