+++ ux.symfony.com/templates/demos/live_component/inline_edit.html.twig
@@ -1,7 +1,7 @@
 {% extends 'demos/live_demo.html.twig' %}
 
 {% block code_block_left %}
-    <twig:CodeBlock filename="src/Twig/InlineEditFood.php" height="400px" />
+    <twig:CodeBlock filename="src/Twig/Components/InlineEditFood.php" height="400px" />
 {% endblock %}
 
 {% block code_block_right %}
