+++ ux.symfony.com/templates/demos/live_component/voting.html.twig
@@ -1,7 +1,7 @@
 {% extends 'demos/live_demo.html.twig' %}
 
 {% block code_block_left %}
-    <twig:CodeBlock filename="src/Twig/FoodVote.php" height="400px" />
+    <twig:CodeBlock filename="src/Twig/Components/FoodVote.php" height="400px" />
 {% endblock %}
 
 {% block code_block_right %}
