+++ ux.symfony.com/templates/ux_packages/twig_component.html.twig
@@ -16,7 +16,7 @@
 {% endblock %}
 
 {% block code_block_left %}
-    <twig:CodeBlock filename="src/Twig/Alert.php" height="330px"/>
+    <twig:CodeBlock filename="src/Twig/Components/Alert.php" height="330px"/>
 {% endblock %}
 
 {% block code_block_right %}
